+++ src/command/preset/preset.ts
@@ -13,7 +13,7 @@ import { findGitRoot } from "../../utils/git";
 export const presetLines = (preset: Preset): string[] => {
     return [
         "",
-        `This repository use the preset "${
+        `The current space use the preset "${
             preset.name
         }" made by ${preset.contributors.join(", ")}: `,
         bold("template: "),