+++ CHANGELOG.md
@@ -40,3 +40,8 @@
 - refactor code
 - add version number to alone command
 - move changelog to his own file
+
+### v1.0.9
+
+- (fix) preset showed a fatal git error if the command was not in a repository
+- (fix) version 1.0.8 due to a functionnal programming error (rightIO(null) instead of rightIO(constVoid) or right(null))